backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-CORS
Flask-SocketIO
psycopg2-binary
cryptography
python-dotenv
PyJWT
bcrypt
requests
langchain
langchain-community
langchain-huggingface
chromadb
pypdf
sentence-transformers
Authlib
pytest
pytest-flask
eventlet
python-multipart
Werkzeug